Move `cloudflare-internal:workers` module to its own impl file.
The code in the new files is purely cut/paste from the old location, no changes.
