+++ src/workerd/api/workers-module.c++
@@ -1,0 +1,61 @@
+// Copyright (c) 2017-2025 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+#include "workers-module.h"
+
+#include <workerd/api/actor-state.h>
+#include <workerd/api/global-scope.h>
+
+namespace workerd::api {
+
+jsg::Ref<WorkerEntrypoint> WorkerEntrypoint::constructor(
+    const v8::FunctionCallbackInfo<v8::Value>& args, jsg::JsObject ctx, jsg::JsObject env) {
+  // HACK: We take `FunctionCallbackInfo` mostly so that we can set properties directly on
+  //   `This()`. There ought to be a better way to get access to `this` in a constructor.
+  //   We *also* declare `ctx` and `env` params more explicitly just for the sake of type checking.
+  jsg::Lock& js = jsg::Lock::from(args.GetIsolate());
+
+  jsg::JsObject self(args.This());
+  self.set(js, "ctx", jsg::JsValue(args[0]));
+  self.set(js, "env", jsg::JsValue(args[1]));
+  return js.alloc<WorkerEntrypoint>();
+}
+
+jsg::Ref<DurableObjectBase> DurableObjectBase::constructor(
+    const v8::FunctionCallbackInfo<v8::Value>& args,
+    jsg::Ref<DurableObjectState> ctx,
+    jsg::JsObject env) {
+  // HACK: We take `FunctionCallbackInfo` mostly so that we can set properties directly on
+  //   `This()`. There ought to be a better way to get access to `this` in a constructor.
+  //   We *also* declare `ctx` and `env` params more explicitly just for the sake of type checking.
+  jsg::Lock& js = jsg::Lock::from(args.GetIsolate());
+
+  jsg::JsObject self(args.This());
+  self.set(js, "ctx", jsg::JsValue(args[0]));
+  self.set(js, "env", jsg::JsValue(args[1]));
+  return js.alloc<DurableObjectBase>();
+}
+
+jsg::Ref<WorkflowEntrypoint> WorkflowEntrypoint::constructor(
+    const v8::FunctionCallbackInfo<v8::Value>& args,
+    jsg::Ref<ExecutionContext> ctx,
+    jsg::JsObject env) {
+  // HACK: We take `FunctionCallbackInfo` mostly so that we can set properties directly on
+  //   `This()`. There ought to be a better way to get access to `this` in a constructor.
+  //   We *also* declare `ctx` and `env` params more explicitly just for the sake of type checking.
+  jsg::Lock& js = jsg::Lock::from(args.GetIsolate());
+
+  jsg::JsObject self(args.This());
+  self.set(js, "ctx", jsg::JsValue(args[0]));
+  self.set(js, "env", jsg::JsValue(args[1]));
+  return js.alloc<WorkflowEntrypoint>();
+}
+
+void EntrypointsModule::waitUntil(kj::Promise<void> promise) {
+  // No need to check if IoContext::hasCurrent since current() will throw
+  // if there is no active request.
+  IoContext::current().addWaitUntil(kj::mv(promise));
+}
+
+}  // namespace workerd::api
