+++ src/workerd/server/workerd-api.c++
@@ -4,8 +4,6 @@
 
 #include "workerd-api.h"
 
-#include "workerd/api/worker-rpc.h"
-
 #include <workerd/api/actor-state.h>
 #include <workerd/api/actor.h>
 #include <workerd/api/analytics-engine.h>
@@ -42,6 +40,8 @@
 #include <workerd/api/urlpattern-standard.h>
 #include <workerd/api/urlpattern.h>
 #include <workerd/api/worker-loader.h>
+#include <workerd/api/worker-rpc.h>
+#include <workerd/api/workers-module.h>
 #include <workerd/io/compatibility-date.h>
 #include <workerd/io/promise-wrapper.h>
 #include <workerd/jsg/jsg.h>
@@ -124,6 +124,7 @@ JSG_DECLARE_ISOLATE_TYPE(JsgWorkerdIsolate,
     EW_EVENTSOURCE_ISOLATE_TYPES,
     EW_WORKER_LOADER_ISOLATE_TYPES,
     EW_MESSAGECHANNEL_ISOLATE_TYPES,
+    EW_WORKERS_MODULE_ISOLATE_TYPES,
     workerd::api::EnvModule,
 
     jsg::TypeWrapperExtension<PromiseWrapper>,
